fix: Verify presence of API KEY

+++ main_script.py
@@ -7,7 +7,7 @@ To use this script, simply run it from the command line:
 $ python main.py
 """
 
-
+import os
 from carbon_bomb import create_carbon_bombs_table
 from connexion import company_involvement_in_carbon_bombs,filter_BOCC_database
 from scrapper import main_scrapping_function, scrapping_company_location
@@ -21,15 +21,19 @@ if __name__ == '__main__':
     # Step2 : Connexion between CarbonBombs and Company
     df = company_involvement_in_carbon_bombs
     df.to_csv("./data_cleaned/connexion_carbonbombs_company.csv",index = False)
-    # Step3 : Scrape bank informations
-    URL = 'https://www.banktrack.org/banks'
-    df = main_scrapping_function(URL)
-    df.to_csv("./data_cleaned/bank_informations.csv",
-              encoding='utf-8-sig',index = False)
-    # Step4 : Scrape company informations
-    df = scrapping_company_location()
-    df.to_csv("./data_cleaned/company_informations.csv",
-              encoding='utf-8-sig',index = False)
+    if os.path.isfile("./credentials.py"):
+        # Step3 : Scrape bank informations
+        URL = 'https://www.banktrack.org/banks'
+        df = main_scrapping_function(URL)
+        df.to_csv("./data_cleaned/bank_informations.csv",
+                encoding='utf-8-sig',index = False)
+        # Step4 : Scrape company informations
+        df = scrapping_company_location()
+        df.to_csv("./data_cleaned/company_informations.csv",
+                encoding='utf-8-sig',index = False)
+    else:
+        print("Create your own Google MAPS API KEY to scrap GPS coordinates"
+              "Skipping Step3 and Step4")
     # Step5 : Connexion between Company and Bank
     df = filter_BOCC_database()
     df.to_csv("./data_cleaned/connexion_bank_company.csv",
